Don't copy recovery.img to BOOTABLE_IMAGES if it doesn't exist
If BOARD_USES_RECOVERY_AS_BOOT is set to true, recovery.img is
not built.

If BOARD_CUSTOM_BOOTIMG is set, INSTALLED_RECOVERYIMAGE_TARGET
gets copied to BOOTABLE_IMAGES.

If both are set, the following error is produced.

acp: missing destination file

Since it is possible to have both BOARD_USES_RECOVERY_AS_BOOT and
BOARD_CUSTOM_BOOTIMG set, guard the copy operation so that it does
not try to copy a missing recovery.img.

Change-Id: Iae24e31a586e98955807d213df3dbaa0a2f13bae
Signed-off-by: Jis G Jacob <studiokeys@blissroms.org>